Fix enum signedness.
In order to obtain a correct (un)signed representation in Python, the signedness of the enum is deduced from the sign of enum values.
Rationale: if there is not any negative value in the enum, then the resulting enum ctype will be unsigned.

Note: It appears that the clang parser is already determining the appropriate sign for each enum values.
We use that information to infer the global type of the enum.

Finally, this fix brings support for 8-byte enum, that were not covered in previous fix 60d796b1cc70b119e87889ce766ef2e45dbf4740